[DDW-665] Storybook story
