+++ storybook/stories/common/Widgets.stories.js
@@ -5,7 +5,6 @@ import { storiesOf } from '@storybook/react';
 import { observable, action as mobxAction } from 'mobx';
 import { action } from '@storybook/addon-actions';
 import { withKnobs, boolean, number, text, date } from '@storybook/addon-knobs';
-import { escapeRegExp } from 'lodash';
 import StoryDecorator from '../_support/StoryDecorator';
 import StoryProvider from '../_support/StoryProvider';
 import StoryLayout from '../_support/StoryLayout';
@@ -21,7 +20,6 @@ import joinSharedIcon from '../../../source/renderer/app/assets/images/join-shar
 import TinySwitch from '../../../source/renderer/app/components/widgets/forms/TinySwitch';
 import ButtonLink from '../../../source/renderer/app/components/widgets/ButtonLink';
 import NormalSwitch from '../../../source/renderer/app/components/widgets/forms/NormalSwitch';
-import ItemsDropdown from '../../../source/renderer/app/components/widgets/forms/ItemsDropdown';
 
 const { intl: enIntl } = new IntlProvider({
   locale: 'en-US',
@@ -113,45 +111,6 @@ storiesOf('Common|Widgets', module)
       format="DD-HH-mm-ss"
     />
   ))
-
-  .add('ItemsDropdown', () => {
-    const options = [
-      {
-        topLabel: 'Item 1 - top',
-        bottomLabel: 'Item 1 - bottom',
-        value: 'item1',
-        label: 'LABEL',
-      },
-      {
-        topLabel: 'Item 2 - top',
-        bottomLabel: 'Item 2 - bottom',
-        value: 'item2',
-      },
-      {
-        topLabel: 'Item 3 - top',
-        bottomLabel: 'Item 3 - bottom',
-        value: 'item3',
-      },
-    ];
-    return (
-      <ItemsDropdown
-        options={options}
-        onChange={action('onChange')}
-        value={'item2'}
-        onSearch={(searchValue) => {
-          return options.filter((option) => {
-            const { topLabel, bottomLabel, value } = option;
-            const regex = new RegExp(escapeRegExp(searchValue), 'i');
-            return (
-              regex.test(topLabel) ||
-              regex.test(bottomLabel) ||
-              regex.test(value)
-            );
-          });
-        }}
-      />
-    );
-  })
 
   .add('InlineEditingInput', () => (
     <div>
